refactor(event): split event tests for GH727
